+++ src/Controllers/item/DeleteController.php
@@ -1,5 +1,6 @@
 <?php
 namespace Controllers\item;
+session_start();
 
 require_once '../../database/DatabaseMysql.php';
 require_once '../../Models/item/Delete.php';
@@ -9,9 +10,14 @@ use Models\item\Delete;
 
 class DeleteController {
     public static function deleteItem() {
+        if (!isset($_SESSION['user']) || isset($_SESSION['user']) && $_SESSION['user']['type'] !== 'admin') {
+            header('location: /items.php');
+            return;
+        }
         $db = new DatabaseMysql;
         $connection = $db->dbConnect();
         Delete::delete($_GET, $connection);
+        header('location: /items.php');
     }
 }
